requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.26.23
botocore==1.29.23
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.4
django-admin-ordering==0.16.1
django-environ==0.9.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-phonenumber-field==7.0.1
django-render-partial==0.4
django-rosetta==0.9.8
django-storages==1.13.1
django-tinymce==3.5.0
django_debug_toolbar==3.8.1
gunicorn==20.1.0
idna==3.4
jdatetime==4.1.0
jmespath==1.0.1
kavenegar==1.1.2
Pillow==9.3.0
polib==1.1.1
python-dateutil==2.8.2
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
urllib3==1.26.13
